% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{app_server}
\alias{app_server}
\title{Server logic for the barnehofte app}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}
}
\value{
A shiny app server object
}
\description{
Server logic for the barnehofte app
}
